+++ activesupport/test/cache/stores/mem_cache_store_test.rb
@@ -297,14 +297,6 @@ class MemCacheStoreTest < ActiveSupport::TestCase
     end
   end
 
-  def test_uses_provided_dalli_client_if_present
-    assert_deprecated(ActiveSupport.deprecator) do
-      host = "custom_host"
-      cache = lookup_store(Dalli::Client.new(host))
-      assert_equal [host], servers(cache)
-    end
-  end
-
   def test_forwards_string_addresses_if_present
     expected_addresses = ["first", "second"]
     cache = lookup_store(expected_addresses)